requirements.txt
setuptools>=61.0
loguru
pyOpenSSL